HGym-Feedback/requirements.txt
gym[atari,accept-rom-license]==0.23.1 # 0.17.2
gym[box2d]==0.23.1
Box2D==2.3.10
atari-py==0.2.6
shortuuid==1.0.1
asyncio==3.4.3
websockets==9.1
numpy==1.18.5
Pillow==8.2.0
boto3==1.14.20
python-dotenv==0.14.0
pyYaml==5.4
gym-super-mario-bros==7.3.2